update dto: align with entity and use decimal

+++ backend/SettlyModels/Dtos/IncomeEmploymentDto.cs
@@ -3,8 +3,8 @@ namespace SettlyModels.Dtos;
 
 public class IncomeEmploymentDto
 {
-    public float MedianIncome { get; set; }
-    public float EmploymentRate { get; set; }
-    public float WhiteCollarRatio { get; set; }
-    public float JobGrowthRate { get; set; }
+    public int MedianIncome { get; set; }
+    public decimal EmploymentRate { get; set; }
+    public decimal WhiteCollarRatio { get; set; }
+    public decimal JobGrowthRate { get; set; }
 }
